+++ backend/src/migrations.rs
@@ -286,7 +286,9 @@ pub enum SchemaVersion {
     AlterPltAccountsAddNotNullConstraint,
     #[display("0043: Add token module pause/unpause status to plt_tokens table")]
     AddTokenModulePauseUnpauseStatus,
-    #[display("0044: Alter PLT tokens add current_supply column")]
+    #[display("0044: Normalize metrics for PLT cumulative transfer amount")]
+    NormalizePltCumulativeTransferAmount,
+    #[display("0045: Alter PLT tokens add current_supply column")]
     AlterPltTokensAddCurrentSupplyColumn,
 }
 impl SchemaVersion {
@@ -361,6 +363,7 @@ impl SchemaVersion {
             SchemaVersion::ReAddBakerRelatedTransactionsIndex => false,
             SchemaVersion::AlterPltAccountsAddNotNullConstraint => false,
             SchemaVersion::AddTokenModulePauseUnpauseStatus => false,
+            SchemaVersion::NormalizePltCumulativeTransferAmount => false,
             SchemaVersion::AlterPltTokensAddCurrentSupplyColumn => false,
         }
     }
@@ -415,6 +418,7 @@ impl SchemaVersion {
             SchemaVersion::ReAddBakerRelatedTransactionsIndex => false,
             SchemaVersion::AlterPltAccountsAddNotNullConstraint => false,
             SchemaVersion::AddTokenModulePauseUnpauseStatus => false,
+            SchemaVersion::NormalizePltCumulativeTransferAmount => false,
             SchemaVersion::AlterPltTokensAddCurrentSupplyColumn => false,
         }
     }
@@ -717,7 +721,15 @@ impl SchemaVersion {
             SchemaVersion::AddTokenModulePauseUnpauseStatus => {
                 tx.as_mut()
                     .execute(sqlx::raw_sql(include_str!(
-                        "./migrations/m0044-alter-plt-tokens-current-supply.sql"
+                        "./migrations/m0044-normalize-metrics-plt-cumulative-transfer-amount.sql"
+                    )))
+                    .await?;
+                SchemaVersion::NormalizePltCumulativeTransferAmount
+            }
+            SchemaVersion::NormalizePltCumulativeTransferAmount => {
+                tx.as_mut()
+                    .execute(sqlx::raw_sql(include_str!(
+                        "./migrations/m0045-alter-plt-tokens-current-supply.sql"
                     )))
                     .await?;
                 SchemaVersion::AlterPltTokensAddCurrentSupplyColumn